+++ src/game/player/elements/Player.tsx
@@ -42,7 +42,7 @@ export default function Player(
     const [state, dispatch] = useReducer(handleAction, new GameState(game, {
         update,
         forceUpdate: () => {
-            (state as GameState).logger.warn("Player", "force update");
+            (state as GameState).logger.weakWarn("Player", "force update");
             flushSync(() => {
                 update();
             });