+++ src/game/player/elements/image/AspectScaleImage.tsx
@@ -24,6 +24,8 @@ export default function AspectScaleImage(
     const {cacheManager} = usePreloaded();
     const {game} = useGame();
 
+    const LogTag = "AspectScaleImage";
+
     function updateWidth() {
         const ref = Ref || imgRef;
         if (ref.current) {
@@ -32,17 +34,19 @@ export default function AspectScaleImage(
     }
 
     useEffect(() => {
-        updateWidth();
-
-        if (props.src && !cacheManager.has(props.src)) {
-            game.getLiveGame().getGameState()?.logger.warn("AspectScaleImage", "Image not preloaded", props.src);
+        if (props.src && (!cacheManager.has(props.src) || cacheManager.isPreloading(props.src))) {
+            game.getLiveGame().getGameState()?.logger.warn(LogTag,
+                `Image not preloaded: "${props.src}". `
+                + "\nThis may be caused by complicated image action behavior that cannot be predicted. "
+                + "\nTo fix this issue, you can manually register the image using scene.requestImagePreload(YourImageSrc). "
+            );
         }
-
-        return ratio.onUpdate(updateWidth);
-    }, [props.src]);
+    }, [props, props.src, id]);
 
     useEffect(() => {
         updateWidth();
+
+        return ratio.onUpdate(updateWidth);
     }, [props, id]);
 
     const src: string = props.src ? (cacheManager.get(props.src) || props.src) : Image.DefaultImagePlaceholder;