+++ src/game/player/elements/preload/Preload.tsx
@@ -1,9 +1,8 @@
-import {useEffect} from "react";
+import {useEffect, useRef} from "react";
 import {GameState} from "@player/gameState";
-import {SrcManager} from "@core/action/srcManager";
+import {ActiveSrc, SrcManager} from "@core/action/srcManager";
 import {usePreloaded} from "@player/provider/preloaded";
 import {Preloaded} from "@player/lib/Preloaded";
-import {PreloadedToken} from "@player/lib/ImageCacheManager";
 import {TaskPool} from "@lib/util/data";
 import {useGame} from "@player/provider/game-state";
 
@@ -14,9 +13,12 @@ export function Preload(
         state: GameState;
     }>) {
     const {preloaded, cacheManager} = usePreloaded();
-    const lastScene = state.getLastScene();
-    const LogTag = "Preload";
     const {game} = useGame();
+    const cachedSrc = useRef<Set<ActiveSrc>>(new Set());
+
+    const LogTag = "Preload";
+    const lastScene = state.getLastScene();
+    const currentAction = game.getLiveGame().getCurrentAction();
 
     /**
      * preload logic 2.0
@@ -25,7 +27,13 @@ export function Preload(
      */
     useEffect(() => {
         if (typeof fetch === "undefined") {
+            preloaded.events.emit(Preloaded.EventTypes["event:preloaded.ready"]);
             state.logger.warn(LogTag, "Fetch is not supported in this environment, skipping preload");
+            return;
+        }
+        if (!game.config.player.preloadAllImages) {
+            preloaded.events.emit(Preloaded.EventTypes["event:preloaded.ready"]);
+            state.logger.debug(LogTag, "Preload all images is disabled, skipping preload");
             return;
         }
         if (game.config.player.forceClearCache) {
@@ -43,7 +51,6 @@ export function Preload(
             game.config.player.preloadDelay,
         );
         const loadedSrc: string[] = [];
-        const tokens: PreloadedToken[] = [];
 
         state.logger.debug(LogTag, "preloading:", sceneSrc);
 
@@ -51,21 +58,21 @@ export function Preload(
             const src = SrcManager.getSrc(image);
             loadedSrc.push(src);
 
-            if (cacheManager.has(src)) {
+            if (cacheManager.has(src) || cacheManager.isPreloading(src)) {
+                state.logger.debug(LogTag, `Image already loaded (${sceneSrc.image.indexOf(image) + 1}/${sceneSrc.image.length})`, src);
                 continue;
             }
             taskPool.addTask(() => new Promise(resolve => {
-                const token = cacheManager.preload(src);
-                token.onFinished(() => {
-                    state.logger.debug(LogTag, `Image loaded (${sceneSrc.image.indexOf(image) + 1}/${sceneSrc.image.length})`, src);
-                    resolve();
-                    tokens.push(token);
-                });
+                cacheManager.preload(src)
+                    .onFinished(() => {
+                        state.logger.debug(LogTag, `Image loaded (${sceneSrc.image.indexOf(image) + 1}/${sceneSrc.image.length})`, src);
+                        resolve();
+                    });
             }));
         }
 
         taskPool.start().then(() => {
-            state.logger.debug(LogTag, "Image preload", `loaded ${cacheManager.size()} images in ${performance.now() - timeStart}ms`);
+            state.logger.info(LogTag, "Image preload", `loaded ${cacheManager.size()} images in ${performance.now() - timeStart}ms`);
 
             if (game.config.player.waitForPreload) {
                 preloaded.events.emit(Preloaded.EventTypes["event:preloaded.ready"]);
@@ -80,14 +87,85 @@ export function Preload(
         preloaded.events.emit(Preloaded.EventTypes["event:preloaded.mount"]);
 
         return () => {
-            for (const token of tokens) {
-                token.abort();
-            }
-
             state.events.emit(GameState.EventTypes["event:state.preload.unmount"]);
             state.logger.debug(LogTag, "Preload unmounted");
         };
     }, [lastScene]);
+
+    /**
+     * Remove cached src when scenes changed
+     */
+    useEffect(() => {
+        cachedSrc.current.clear();
+    }, [lastScene]);
+
+    /**
+     * predict preload logic
+     *
+     * Get future src and preload them
+     */
+    useEffect(() => {
+        if (typeof fetch === "undefined") {
+            return;
+        }
+        if (game.config.player.preloadAllImages) {
+            return;
+        }
+
+        const timeStart = performance.now();
+        const allSrc: ActiveSrc[] = game
+            .getLiveGame()
+            .getAllPredictableActions(currentAction, game.config.player.maxPreloadActions)
+            .map(s => SrcManager.getPreloadableSrc(s))
+            .filter<ActiveSrc>(function (src): src is ActiveSrc {
+                return src !== null;
+            });
+        const sceneBasedSrc =
+            allSrc.filter(function (src): src is ActiveSrc<"scene"> {
+                return src?.activeType === "scene";
+            });
+        sceneBasedSrc.forEach(src => {
+            if (cachedSrc.current.has(src)) {
+                return;
+            }
+            cachedSrc.current.add(src);
+        });
+
+        const actionSrc = SrcManager.catSrc([
+            ...cachedSrc.current,
+            ...allSrc,
+        ]);
+
+        const taskPool = new TaskPool(
+            game.config.player.preloadConcurrency,
+            game.config.player.preloadDelay,
+        );
+        const preloadSrc: string[] = [];
+
+        state.logger.debug(LogTag, "preloading:", actionSrc);
+
+        for (const image of actionSrc.image) {
+            const src = SrcManager.getSrc(image);
+            preloadSrc.push(src);
+
+            if (cacheManager.has(src) || cacheManager.isPreloading(src)) {
+                state.logger.debug(LogTag, `Image already loaded (${actionSrc.image.indexOf(image) + 1}/${actionSrc.image.length})`, src);
+                continue;
+            }
+            taskPool.addTask(() => new Promise(resolve => {
+                cacheManager.preload(src)
+                    .onFinished(() => {
+                        state.logger.debug(LogTag, `Image loaded (${actionSrc.image.indexOf(image) + 1}/${actionSrc.image.length})`, src);
+                        resolve();
+                    });
+            }));
+        }
+
+        taskPool.start().then(() => {
+            state.logger.info(LogTag, "Image preload (quick reload)", `loaded ${cacheManager.size()} images in ${performance.now() - timeStart}ms`);
+            cacheManager.filter(preloadSrc);
+        });
+    }, [currentAction]);
 
     return null;
 }